Refactored the `ShowcaseForm` component to use the `ItemApi` helper instead of the `ShowcaseApi`, and updated the `ShowcaseItem` component with a more visually appealing layout.
