+++ src/Components/Showcase/ShowcaseForm.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { fetchData } from "../../helpers/ShowcaseApi";
+import { fetchData } from "../../helpers/ItemApi";
 import { createItem, searchItem } from "../../helpers/itemHelper";
 import { useAuth } from "../../helpers/AuthContext";
 import data from '../../mockData/data.json'
@@ -71,10 +71,10 @@ export default function ShowcaseForm() {
         <button type="submit">Submit</button>
       </form>
       <br />
-      <h2>Item Details:</h2>
+      {/* <h2>Item Details:</h2>
       {
         JSON.stringify(items, null, 2)
-      }
+      } */}
       <hr />
       <ShowcaseList items={items} />
     </>
