+++ src/helpers/itemHelper.jsx
@@ -1,72 +1,67 @@
 import axios from 'axios';
 
-class ItemController {
-  constructor() {
-    this.API_URL = 'http://localhost:3000/items';
-  }
+const API_URL = 'http://localhost:3000/items';
 
-  async getAllItems() {
-    try {
-      const response = await axios.get(this.API_URL);
-      return response.data;
-    } catch (error) {
-      console.error('Error fetching items:', error);
-      throw error;
+export const getAllItems = async () => {
+  try {
+    const response = await axios.get(API_URL);
+    return response.data;
+  } catch (error) {
+    console.error('Error fetching items:', error);
+    throw error;
+  }
+};
+
+export const createItem = async (uid, productEan) => {
+  try {
+    const response = await axios.post(API_URL, { uid, productEan });
+    return response.data;
+  } catch (error) {
+    console.error('Error creating item:', error);
+    throw error;
+  }
+};
+
+export const updateItemById = async (id, name, description) => {
+  try {
+    const response = await axios.put(`${API_URL}/${id}`, { name, description });
+    return response.data;
+  } catch (error) {
+    console.error('Error updating item:', error);
+    throw error;
+  }
+};
+
+export const deleteItemById = async (id) => {
+  try {
+    const response = await axios.delete(`${API_URL}/${id}`);
+    return response.data;
+  } catch (error) {
+    console.error('Error deleting item:', error);
+    throw error;
+  }
+};
+
+// Note: These functions are not present in the provided controller, but you may want to keep them if they're used elsewhere
+export const getItemById = async (id) => {
+  try {
+    const response = await axios.get(`${API_URL}/${id}`);
+    return response.data;
+  } catch (error) {
+    console.error('Error fetching item:', error);
+    throw error;
+  }
+};
+
+export const searchItem = async (code, type) => {
+  try {
+    const response = await axios.get(`${API_URL}/search`, { params: { code, type } });
+    return response.data;
+  } catch (error) {
+    if (error.response && error.response.status === 404) {
+      return null;
     }
+    console.error('Error searching for item:', error);
+    throw error;
   }
-
-  async getItemById(id) {
-    try {
-      const response = await axios.get(`${this.API_URL}/${id}`);
-      return response.data;
-    } catch (error) {
-      console.error('Error fetching item:', error);
-      throw error;
-    }
-  }
-
-  async createItem(itemData, uid) {
-    try {
-      const response = await axios.post(this.API_URL, { ...itemData, uid });
-      return response.data;
-    } catch (error) {
-      console.error('Error creating item:', error);
-      throw error;
-    }
-  }
-
-  async updateItemById(id, itemData) {
-    try {
-      const response = await axios.put(`${this.API_URL}/${id}`, itemData);
-      return response.data;
-    } catch (error) {
-      console.error('Error updating item:', error);
-      throw error;
-    }
-  }
-
-  async deleteItemById(id) {
-    try {
-      const response = await axios.delete(`${this.API_URL}/${id}`);
-      return response.data;
-    } catch (error) {
-      console.error('Error deleting item:', error);
-      throw error;
-    }
-  }
-
-  async searchItem(code, type) {
-    try {
-      const response = await axios.get(`${this.API_URL}/search`, { params: { code, type } });
-      return response.data;
-    } catch (error) {
-      if (error.response && error.response.status === 404) {
-        return null;
-      }
-      console.error('Error searching for item:', error);
-      throw error;
-    }
-  }
-}
-
-export default new ItemController();
+};
